task-1/requirements.txt
numpy==1.23.5
Pillow==9.5.0
Pillow==11.1.0
scikit_learn==1.6.1
tensorflow==2.12.0
tensorflow_intel==2.12.0
